IOS-7049 Rename network provider class, add pending transaction reset, move balance request creation inside network service
